sfeed: prefer published over updated
It was useful to kept track of updated news articles for example, sadly
some sites (like youtube) bump the "updated" timestamp everytime a user
likes a video or farts. This makes it totally useless and messes
up the sort order of "news".

Now prefer "published" over "updated" for Atom feeds.