sfeed_update: fix time modified on feed first run
A feeds file was created with touch when it didnt exist. This confuses
curl -z since it uses that file for If-Modified-Since. Now it just
copies the temporary file to the new file without merging if it doesn't
exist. The warning for curl for a non-existant file is already suppressed
with curl -s.