sfeed_update: fix issue with merging failed feeds
When fetching a feed failed and its temporary file was stored (filesize=0)
it would still be merged with the (possible) old file. This updated the
modification time which would be used in the next poll (If-Modified-Since).

The solution is to check if the file is non-empty and only then merge, this
is still not 100% fool-proof, but much better.