sfeed_frames: overhaul
sfeed_frames used to write HTML pages for each entry for each feed. This can
be useful but had security issues, because the context of the content changes.

sfeed_frames is now a HTML version which works better with browsers that don't
support CSS or tables well like w3m and lynx. It is now an alternative for
sfeed_html.

- Don't reference and embed HTML content for security reasons. This was
  documented under "SECURITY CONSIDERATIONS" in the man page.
- Tighten pledge(2).
- Simplify