stricter Atom link parsing
the Atom link parsing is more strict now and checks the rel attribute. When the
rel attribute is empty it is handled as a normal link ("alternate").

This makes sure when an link with an other type is specified (such as
"enclosure", "related", "self" or "via") before a link it is not used.

sfeed does not handle enclosures, but the code is reworked so it is very simple
to add this. Enclosures are often used for example to attach some image to a
newspost or an audio file to a podcast.